Started customizing vim configuration options.
I added two color schemes to test with VIM.  So far I've only used the molokai
scheme.  They need to be edited to change the colors for invisible characters.  The
vimrc file is not done yet, either.  There are still several options needed and I
found a site to serve as an example.  It is already starting to look really good.